requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docutils==0.20.1
idna==3.6
Kivy==2.3.0
Kivy-Garden==0.1.5
kivymd==1.2.0
pillow==10.2.0
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
